% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{characteristic_curve}
\alias{characteristic_curve}
\title{Internal characteristic curve function}
\usage{
characteristic_curve(item, fit, range, draws, seed, FUN)
}
\arguments{
\item{item}{item indicator (an integer)}

\item{fit}{posterior samples from an IRT model}

\item{range}{x axis limits}

\item{draws}{number of random parameter draws from the posterior}

\item{seed}{sets the seed for random draws of the posterior replications}

\item{FUN}{declares the function to be used to calculate curves (see details)}
}
\description{
plots characteristic and information curves
}
\details{
Supported functions are (1) icf, which plots item/test characteristic functions, and 
(2) iif, which plots item/test information curves.
}
