% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\title{item characteristic curve}
\usage{
icc(item, fit, range = 5, draws = 0, seed = NULL)
}
\arguments{
\item{item}{item indicator (an integer)}

\item{fit}{posterior samples from an IRT model}

\item{range}{x axis limits}

\item{draws}{number of random parameter draws from the posterior}
}
\description{
plots the item characteristic curve for a given item
}
