% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icf}
\alias{icf}
\title{item characteristic function}
\usage{
icf(alpha, beta, theta)
}
\arguments{
\item{alpha}{item discrimination}

\item{beta}{item difficulty}

\item{theta}{person parameter}
}
\description{
calculates the value of the item characteristic function
}
