% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_stanlist.R
\name{long_to_stanlist}
\alias{long_to_stanlist}
\title{Stan-List from a long data.frame}
\usage{
long_to_stanlist(data, measure.var, item.id, person.id, group.id = NULL)
}
\arguments{
\item{data}{A data.frame in long format}

\item{measure.var}{Variable containing item responses}

\item{item.id}{Variable containing item identifiers}

\item{person.id}{Variable containing person identifiers}

\item{group.id}{Variable containing group identifiers (optional)}
}
\description{
Transfroms the data to an appropriate list for usage with Stan IRT Models
}
