% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictive_checks.R
\name{ppc_observed_score}
\alias{ppc_observed_score}
\title{Posterior predictive score distribution plot}
\usage{
ppc_observed_score(y_rep, data, draws = NULL, seed = NULL, alpha = 0.1)
}
\arguments{
\item{y_rep}{posterior replications of a stanfit object}

\item{data}{Stan data file (named list)}

\item{draws}{number of random draws from y_rep}

\item{seed}{set a seed for the random number generator}

\item{alpha}{opacity of lines for posterior predicted scores}
}
\description{
Posterior predictive check of observed score distribution for IRT models
}
