% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_functions.R
\name{stan_1pl}
\alias{stan_1pl}
\title{Fitting function for the 1PL-IRT Model}
\usage{
stan_1pl(data, ...)
}
\arguments{
\item{data}{A named list containing all relevant variables (see details)}
}
\description{
Fits the 1PL Item Response Model
}
\details{
The named list must contain the following variables:
}
