% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{tcc}
\alias{tcc}
\title{Test characteristic curve function}
\usage{
tcc(fit, range = 3, draws = NULL, seed = NULL, alpha = 0.1)
}
\arguments{
\item{fit}{posterior samples from an IRT model}

\item{range}{x axis limits}

\item{draws}{number of random parameter draws from the posterior}

\item{seed}{sets the seed for random draws of posterior replications}
}
\description{
plots test characteristic curves
}
\details{
High level wrapper for \code{characteristic_curve} used for plotting test characteristic 
curves.
}
