% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_stanlist.R
\name{wide_to_stanlist}
\alias{wide_to_stanlist}
\title{Stan-List from a wide data.frame}
\usage{
wide_to_stanlist(data, measure.vars, group.var = NULL)
}
\arguments{
\item{data}{A data.frame in wide format}

\item{measure.vars}{character string containing column names of item responses}

\item{group.var}{character string containing a grouping variable name (optional)}
}
\description{
Transfroms the data to an appropriate list for usage with Stan IRT Models
}
